Move base join building logic to SqlQueryPlanJoinBuilder
Our base join logic is fairly involved, but it's also self-contained.
This reduces overhead in reading the DataflowToSqlQueryPlanConverter
by moving that logic to a dedicated location.

This is a fairly strict cut/paste, additional cleanup will streamline
the logic within the sql_join_builder module.
